SQL commands using sql magic extension.
This project was done for IBM Certificate "Databases and SQL for Data Science". The SQL Commands were run using SQL Magic extension used in JupyterLab notebook. We will connect to a Db2 database, issue SQL commands to create tables, insert data, and run queries, as well as retrieve results in a Python dataframe.  The url below contains the IBM Badge for successful completion of project and course.
https://www.youracclaim.com/badges/58c81c53-3cc8-46c1-aa91-e7fbcb8beb37/public_url